requirements.txt
asgiref==3.7.2
Django==3.2.20
django-mysql==4.10.0
Pillow==9.5.0
PyMySQL==1.1.0
pytz==2023.3
sqlparse==0.4.4
django-cors-headers==4.1.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
whitenoise==6.5.0
django-ckeditor==6.3.2
django-admin-thumbnails==0.2.8